added action getUser, fix filtering by type, adding scroll to dashboard space

+++ server/services/spaceService.js
@@ -49,10 +49,14 @@ module.exports = {
     BlogRepository.create({})
       .then(blog => {
         const spaceWithOwnerAndEmptyBlog = { ...req.body, ownerId: req.user._id, blogId: blog._id }
+        console.log(spaceWithOwnerAndEmptyBlog.ownerId)
         SpaceRepository.create(spaceWithOwnerAndEmptyBlog)
           .then(space => {
-            UserRepository.addSpaceToUser({userId: req.user._id, spaceId: space._id})
-              .then(() => res.json(space))
+            UserRepository.addSpaceToUser({userId: spaceWithOwnerAndEmptyBlog.ownerId, spaceId: space._id})
+              .then((user) => {
+                console.log(user)
+                return res.json(space)
+              })
               .catch(err => console.log(err))
           })
           .catch((err) => {
@@ -95,8 +99,29 @@ module.exports = {
       return res.end('Invalid id')
     }
 
-    SpaceRepository.delete(id)
-      .then(data => res.json(data))
+    SpaceRepository.getById(id)
+      .then((data) => {
+        if (data.length === 0) {
+          res.status(404)
+          return res.end()
+        }
+        UserRepository.getById(data[0].owner._id)
+          .then(user => {
+            if (data[0].owner._id.equals(user._id)) {
+              UserRepository.deleteSpace(user._id, data[0]._id)
+                .then(user => console.log(user))
+                .catch(err => console.log(err))
+            }
+            SpaceRepository.delete(id)
+              .then(data => res.json(data))
+              .catch((err) => {
+                console.log(err)
+                res.status(400)
+                res.end()
+              })
+          })
+          .catch(err => console.log(err))
+      })
       .catch((err) => {
         console.log(err)
         res.status(400)
